+++ src/pages/Dashboard.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import { getCampaigns } from '../reducers/campaignsReducer';
+import { getCampaigns } from '../reducers/campaign';
 
 export default function Dashboard() {
   const dispatch = useDispatch();
